feat: close integration execution loop

+++ onelake-app/module-integration/src/main/java/com/onelake/integration/dto/SyncRunDTO.java
@@ -14,7 +14,7 @@ public record SyncRunDTO(
     String errorMsg,
     String checkpoint,
     Long durationMs,
-    Long throughputRows,
+    Double throughputRows,
     Instant startedAt,
     Instant finishedAt
 ) {}
